pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "credential-downscope"
version = "0.1.0"
description = "Least-privilege credential injection for CLI tools and MCP servers"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.21.1",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.scripts]
credential-downscope-proxy = "credential_downscope.mcp_proxy:main"

[tool.hatch.build.targets.wheel]
packages = ["src/credential_downscope"]

[tool.hatch.build]
sources = ["src"]
